lint

+++ src/lib/server/repos/oauth.repo.ts
@@ -1,14 +1,17 @@
 import { and, eq } from 'drizzle-orm'
 import { db } from '../db'
-import { oauths, type OauthDb, type OauthDbCreate, type OauthProvider, type UserId } from '../db/schema'
+import {
+	oauths,
+	type OauthDb,
+	type OauthDbCreate,
+	type OauthProvider,
+	type UserId
+} from '../db/schema'
 
 export async function createOauth(create: OauthDbCreate) {
-  return (await db.insert(oauths).values(create).returning())[0]
+	return (await db.insert(oauths).values(create).returning())[0]
 }
 
 export async function deleteOauth(userId: UserId, provider: OauthProvider) {
-  await db.delete(oauths).where(and(
-    eq(oauths.userId, userId),
-    eq(oauths.provider, provider)
-  ))
+	await db.delete(oauths).where(and(eq(oauths.userId, userId), eq(oauths.provider, provider)))
 }
